requirements.txt
asgiref==3.3.0
Django==3.1.3
django-ckeditor==6.0.0
django-crispy-forms==1.9.2
django-filter==2.4.0
django-js-asset==1.2.2
djangorestframework==3.12.1
Pillow==8.0.1
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pypng==0.0.20
PyQRCode==1.2.1
pytz==2020.4
sqlparse==0.4.1
